refactoring

+++ HydraListAndroid/src/com/fada21/android/hydralist/util/HydraListConsts.java
@@ -2,6 +2,6 @@ package com.fada21.android.hydralist.util;
 
 public class HydraListConsts {
 
-    public static final int INVALID_ID = -1;
+    public static final long INVALID_ID = -1;
     public static final int UNSET   = -1;
 }
